Allow multiple namespaces deep
In the code context parsing, the general method parser currently allows for multiple namespaces, e.g. `NS1.NS2.myMethod = ...` through `([\w$.]+)`, but the prototype method and property regexes don't allow this. They instead have `([\w$.]+)`.

Not sure if there's a reason for this, but this PR is a simple patch to make the other regex like the general method one.